+++ source/isaaclab_physx/isaaclab_physx/assets/rigid_object/rigid_object.py
@@ -16,7 +16,6 @@ import warp as wp
 from pxr import UsdPhysics
 
 import isaaclab.sim as sim_utils
-import isaaclab.utils.string as string_utils
 from isaaclab.assets.rigid_object.base_rigid_object import BaseRigidObject
 from isaaclab.utils.wrench_composer import WrenchComposer
 
@@ -189,7 +188,7 @@ class RigidObject(BaseRigidObject):
         Returns:
             A tuple of lists containing the body indices and names.
         """
-        return string_utils.resolve_matching_names(name_keys, self.body_names, preserve_order)
+        return self._resolve_matching_names_cached(name_keys, self.body_names, preserve_order)
 
     """
     Operations - Write to simulation.